Revert "Teach the build system about mod_dash"
This reverts commit b1a3892a5b..., because I found a more effective
way to wire in mod_dash, that is, I include it as a folder inside the
root directory, and I patch the system path accordingly.

To wire in mod_dash with the current diff, instead, I would need to
add a special case for the zipimport in utils_modules.py, which seems
to me less robust on the long run.