add 0001.patch: include ROOTDIR in sys.path if needed
Py2exe configures the system path to include the zipfile only.

We ship the mod_dash as a folder outside of the zipfile, because
this kind of deployment fits better with utils_modules.py code.

Therefore we also need to patch the system path to include the
root directory, which contains the mod_dash.